+++ source/robot_model/include/robot_model/Model.hpp
@@ -17,13 +17,13 @@ using namespace std::chrono_literals;
 
 namespace robot_model {
 /**
- * @brief parameters for the inverse geometry function
+ * @brief parameters for the inverse kinematics function
  * @param damp damping added to the diagonal of J*Jt in order to avoid the singularity
  * @param alpha alpha €]0,1], it is used to make Newthon-Raphson method less aggressive
  * @param gamma gamma €]0,1], represents the strength of the repulsive potential field in the Clamping Weighted Least-Norm method
- * @param margin the distance from the joint limit on which we want to penalize the joint position
+ * @param margin the distance from the joint limit at which the joint positions should be penalized (rad)
  * @param tolerance the maximum error tolerated between the desired cartesian state and the one obtained by the returned joint positions
- * @param max_number_of_iterations the maximum number of iterations that the algorithm do for solving the inverse geometry
+ * @param max_number_of_iterations the maximum number of iterations that the algorithm do for solving the inverse kinematics
  */
 struct InverseKinematicsParameters {
   double damp = 1e-6;
@@ -81,6 +81,20 @@ private:
   bool init_qp_solver();
 
   /**
+   * @brief Check if frames exist in robot model and return its ids
+   * @param frame_names containing the frame names to check
+   * @return the ids of the frames
+   */
+  std::vector<unsigned int> get_frame_ids(const std::vector<std::string>& frame_names);
+
+  /**
+   * @brief Check if a frame exist in robot model and return its id
+   * @param frame_name containing the frame name to check
+   * @return the id of the frame if it exists
+   */
+  unsigned int get_frame_id(const std::string& frame_name);
+
+  /**
    * @brief Compute the Jacobian from given joint positions at the frame in parameter
    * @param joint_positions containing the joint positions of the robot
    * @param joint_id id of the frame at which to compute the Jacobian
@@ -101,18 +115,18 @@ private:
                                                    unsigned int frame_id);
 
   /**
-   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint values
+   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint positions
    * @param joint_positions the joint state of the robot
-   * @param frame_ids ids of the frames at which we want to extract the pose
+   * @param frame_ids ids of the frames at which to extract the pose
    * @return the desired poses
    */
   std::vector<state_representation::CartesianPose> forward_kinematics(const state_representation::JointPositions& joint_positions,
                                                                       const std::vector<unsigned int>& frame_ids);
 
   /**
-   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint values for a single frame
+   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint positions for a single frame
    * @param joint_positions the joint state of the robot
-   * @param frame_id id of the frames at which we want to extract the pose
+   * @param frame_id id of the frames at which to extract the pose
    * @return the desired pose
    */
   state_representation::CartesianPose forward_kinematics(const state_representation::JointPositions& joint_positions,
@@ -142,16 +156,16 @@ private:
 
   /**
    * @brief Compute the weighted matrix of the algorithm "Clamping Weighted Least-Norm"
-   * @param joint_positions the joint position at the current iteration in the inverse geometry problem
-   * @param margin the distance from the joint limit on which we want to penalize the joint position
+   * @param joint_positions the joint position at the current iteration in the inverse kinematics problem
+   * @param margin the distance from the joint limit at which the joint positions should be penalized
    * @return the weighted matrix
    */
   Eigen::MatrixXd cwln_weighted_matrix(const state_representation::JointPositions& joint_positions, double margin);
 
   /**
    * @brief Compute the repulsive potential field of the algorithm "Clamping Weighted Least-Norm"
-   * @param joint_positions the joint position at the current iteration in the inverse geometry problem
-   * @param margin the distance from the joint limit on which we want to penalize the joint position
+   * @param joint_positions the joint position at the current iteration in the inverse kinematics problem
+   * @param margin the distance from the joint limit at which the joint positions should be penalized
    * @return the repulsive potential field
    */
   Eigen::VectorXd cwln_repulsive_potential_field(const state_representation::JointPositions& joint_positions,
@@ -160,8 +174,8 @@ private:
   /**
    * @brief Check the arguments of the inverse_velocity function and throw exceptions if they are not correct
    * @param cartesian_twists vector of twist
-   * @param joint_positions current joint positions, used to compute the jacobian matrix
-   * @param frame_names names of the frames at which we want to compute the twists
+   * @param joint_positions current joint positions, used to compute the Jacobian matrix
+   * @param frame_names names of the frames at which to compute the twists
    */
   void check_inverse_velocity_arguments(const std::vector<state_representation::CartesianTwist>& cartesian_twists,
                                         const state_representation::JointPositions& joint_positions,
@@ -264,7 +278,7 @@ public:
 
   /**
    * @brief Compute the Jacobian from a given joint state at the frame given in parameter
-   * @param joint_positions containing the joint values of the robot
+   * @param joint_positions containing the joint positions of the robot
    * @param frame_name name of the frame at which to compute the Jacobian, if empty computed for the last frame
    * @return the Jacobian matrix
    */
@@ -320,29 +334,29 @@ public:
   state_representation::JointTorques compute_gravity_torques(const state_representation::JointPositions& joint_positions);
 
   /**
-   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint values
+   * @brief Compute the forward kinematics, i.e. the pose of certain frames from the joint positions
    * @param joint_positions the joint state of the robot
-   * @param frame_names names of the frames at which we want to extract the pose
+   * @param frame_names names of the frames at which to extract the poses
    * @return the pose of desired frames
    */
   std::vector<state_representation::CartesianPose> forward_kinematics(const state_representation::JointPositions& joint_positions,
                                                                       const std::vector<std::string>& frame_names);
 
   /**
-   * @brief Compute the forward kinematics, i.e. the pose of the frame from the joint values
+   * @brief Compute the forward kinematics, i.e. the pose of the frame from the joint positions
    * @param joint_positions the joint state of the robot
-   * @param frame_name name of the frame at which we want to extract the pose
+   * @param frame_name name of the frame at which to extract the pose
    * @return the pose of the desired frame
    */
   state_representation::CartesianPose forward_kinematics(const state_representation::JointPositions& joint_positions,
-                                                         std::string frame_name = "");
+                                                         const std::string& frame_name = "");
 
   /**
-   * @brief Compute the inverse geometry, i.e. joint values from the pose of the end-effector in a iteratively manner
+   * @brief Compute the inverse kinematics, i.e. joint positions from the pose of the end-effector in an iterative manner
    * @param cartesian_pose containing the desired pose of the end-effector
-   * @param parameters parameters of the inverse geometry algorithm (default is default values of the
+   * @param parameters parameters of the inverse kinematics algorithm (default is default values of the
    * InverseKinematicsParameters structure)
-   * @param frame_name name of the frame at which we want to extract the pose
+   * @param frame_name name of the frame at which to extract the pose
    * @return the joint positions of the robot
    */
   state_representation::JointPositions inverse_kinematics(const state_representation::CartesianPose& cartesian_pose,
@@ -350,12 +364,12 @@ public:
                                                           const std::string& frame_name = "");
 
   /**
-   * @brief Compute the inverse kinematics, i.e. joint values from the pose of the end-effector
+   * @brief Compute the inverse kinematics, i.e. joint positions from the pose of the end-effector
    * @param cartesian_pose containing the desired pose of the end-effector
    * @param joint_positions current state of the robot containing the generalized position
-   * @param parameters parameters of the inverse geometry algorithm (default is default values of the
+   * @param parameters parameters of the inverse kinematics algorithm (default is default values of the
    * InverseKinematicsParameters structure)
-   * @param frame_name name of the frame at which we want to extract the pose
+   * @param frame_name name of the frame at which to extract the pose
    * @return the joint positions of the robot
    */
   state_representation::JointPositions inverse_kinematics(const state_representation::CartesianPose& cartesian_pose,
@@ -364,9 +378,9 @@ public:
                                                           const std::string& frame_name = "");
 
   /**
-   * @brief Compute the forward velocity kinematics, i.e. the twist of certain frames from the joint values
+   * @brief Compute the forward velocity kinematics, i.e. the twist of certain frames from the joint states
    * @param joint_state the joint state of the robot with positions to compute the Jacobian and velocities for the twist
-   * @param frame_names name of the frames at which we want to compute the twist
+   * @param frame_names name of the frames at which to compute the twist
    * @return the twists of the frames in parameter
    */
   std::vector<state_representation::CartesianTwist> forward_velocity(const state_representation::JointState& joint_state,
@@ -375,7 +389,7 @@ public:
   /**
    * @brief Compute the forward velocity kinematics, i.e. the twist of the end-effector from the joint velocities
    * @param joint_state the joint state of the robot with positions to compute the Jacobian and velocities for the twist
-   * @param frame_name name of the frame at which we want to compute the twist
+   * @param frame_name name of the frame at which to compute the twist
    * @return the twist of the frame in parameter
    */
   state_representation::CartesianTwist forward_velocity(const state_representation::JointState& joint_state,
@@ -385,8 +399,8 @@ public:
    * @brief Compute the inverse velocity kinematics, i.e. joint velocities from the velocities of the frames in parameter
    * using the Jacobian
    * @param cartesian_twists vector of twist
-   * @param joint_positions current joint positions, used to compute the jacobian matrix
-   * @param frame_names names of the frames at which we want to compute the twists
+   * @param joint_positions current joint positions, used to compute the Jacobian matrix
+   * @param frame_names names of the frames at which to compute the twists
    * @return the joint velocities of the robot
    */
   state_representation::JointVelocities inverse_velocity(const std::vector<state_representation::CartesianTwist>& cartesian_twists,
@@ -397,8 +411,8 @@ public:
    * @brief Compute the inverse velocity kinematics, i.e. joint velocities from the twist of the end-effector using the
    * Jacobian
    * @param cartesian_twist containing the twist of the end-effector
-   * @param joint_positions current joint positions, used to compute the jacobian matrix
-   * @param frame_name name of the frame at which we want to compute the twist
+   * @param joint_positions current joint positions, used to compute the Jacobian matrix
+   * @param frame_name name of the frame at which to compute the twist
    * @param parameters parameters of the inverse velocity kinematics algorithm (default is default values of the
    * QPInverseVelocityParameters structure)
    * @return the joint velocities of the robot
@@ -414,7 +428,7 @@ public:
    * @param joint_positions current joint positions, used to compute the jacobian matrix
    * @param parameters parameters of the inverse velocity kinematics algorithm (default is default values of the
    * QPInverseVelocityParameters structure)
-   * @param frame_names names of the frames at which we want to compute the twists
+   * @param frame_names names of the frames at which to compute the twists
    * @return the joint velocities of the robot
    */
   state_representation::JointVelocities inverse_velocity(const std::vector<state_representation::CartesianTwist>& cartesian_twists,
@@ -427,10 +441,10 @@ public:
    * QP optimization method
    * @param cartesian_twist containing the twist of the end-effector
    * @param joint_positions current joint positions, used to compute the Jacobian matrix
-   * @param joint_positions current joint positions, used to compute the jacobian matrix
+   * @param joint_positions current joint positions, used to compute the Jacobian matrix
    * @param parameters parameters of the inverse velocity kinematics algorithm (default is default values of the
    * QPInverseVelocityParameters structure)
-   * @param frame_name name of the frame at which we want to compute the twist
+   * @param frame_name name of the frame at which to compute the twist
    * @return the joint velocities of the robot
    */
   state_representation::JointVelocities inverse_velocity(const state_representation::CartesianTwist& cartesian_twist,